Remove AbstractOriginSource (#1019)
